feat(api): enhance user authentication and chatbot integration
- Implemented password change functionality in AuthController with validation for current and new passwords.
- Updated ChatController to include user ID in message handling and history retrieval, ensuring session access control.
- Enhanced error handling for unauthorized access and weak password scenarios.
- Adjusted DTOs and models to support new password change feature and improved data handling in chat operations.
- Refactored services to accommodate user-specific chat session management and retention.

+++ Api/Dtos/DashboardDtos.cs
@@ -17,8 +17,8 @@ public sealed class LowStockItemDto
     public Guid Id { get; set; }
     public string Name { get; set; } = string.Empty;
     public string Sku { get; set; } = string.Empty;
-    public int CurrentStock { get; set; }
-    public int ReorderLevel { get; set; }
+    public decimal CurrentStock { get; set; }
+    public decimal ReorderLevel { get; set; }
     public string Status { get; set; } = "low_stock";
 }
