+++ Domain/Entities/Inventory.cs
@@ -7,9 +7,9 @@ public class Inventory
     public Product Product { get; set; } = null!;
     public Guid WarehouseId { get; set; }
     public Warehouse Warehouse { get; set; } = null!;
-    public int CurrentStock { get; set; }
-    public int MinStock { get; set; }
-    public int MaxStock { get; set; }
+    public decimal CurrentStock { get; set; }
+    public decimal MinStock { get; set; }
+    public decimal MaxStock { get; set; }
     public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
 
     public ICollection<InventoryMovement> Movements { get; set; } = [];
